requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
SQLAlchemy==2.0.32
APScheduler==3.10.4
httpx==0.27.2
nest-asyncio==1.6.0
